+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.14.5)
+
+👌 **Improvements**
+
+- alphabet is now checked for duplicate characters in `BigInteger::(from|to)ArbitraryBase()`
+
 ## [0.14.4](https://github.com/brick/math/releases/tag/0.14.4) - 2026-02-02
 
 🗑️ **Deprecations**